Port dev controller modules to new architecture
